+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1
 
 - Add support for libraries.
+- Add support for abstract contracts and interfaces.
 
 ## 0.1.0